Create GoToTopButton component
