+++ src/app/content/components/GoToTopButton.tsx
@@ -1,0 +1,61 @@
+import React from 'react';
+import { FormattedMessage } from 'react-intl';
+import styled from 'styled-components';
+import { AngleUp } from 'styled-icons/fa-solid/AngleUp';
+
+// tslint:disable-next-line:variable-name
+export const GoToTopWrapper = styled.div`
+  width: 4.8rem;
+  height: 4.8rem;
+  position: absolute;
+  z-index: 2;
+  bottom: 4.8rem;
+  right: 4.8rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+`;
+
+// tslint:disable-next-line:variable-name
+export const GoToTop = styled.div`
+  width: 2.4rem;
+  height: 2.4rem;
+  background: #959595;
+  border-radius: 50%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: white;
+`;
+
+// tslint:disable-next-line:variable-name
+export const GoToTopIcon = styled(AngleUp)`
+  width: 1.6rem;
+  height: 1.6rem;
+`;
+
+interface GoToTopButtonProps {
+  i18nAriaLabel: string;
+  onClick: () => void;
+  [key: string]: any;
+}
+
+// tslint:disable-next-line: variable-name
+const GoToTopButton = ({ i18nAriaLabel, onClick, ...rest }: GoToTopButtonProps) => (
+  <FormattedMessage id={i18nAriaLabel}>
+    {(msg: string) => (
+      <GoToTopWrapper
+        onClick={onClick}
+        aria-label={msg}
+        {...rest}
+      >
+        <GoToTop>
+          <GoToTopIcon />
+        </GoToTop>
+      </GoToTopWrapper>
+    )}
+  </FormattedMessage>
+);
+
+export default GoToTopButton;
