+++ src/app/content/highlights/components/ShowMyHighlights.tsx
@@ -1,11 +1,11 @@
 import { HTMLElement } from '@openstax/types/lib.dom';
 import flow from 'lodash/fp/flow';
 import React, { Component } from 'react';
-import { FormattedMessage } from 'react-intl';
 import { connect } from 'react-redux';
 import withServices from '../../../context/Services';
 import { isHtmlElement } from '../../../guards';
 import { AppServices, AppState, Dispatch } from '../../../types';
+import GoToTopButton from '../../components/GoToTopButton';
 import { loadMoreSummaryHighlights } from '../actions';
 import { loadMoreDistanceFromBottom } from '../constants';
 import * as select from '../selectors';
@@ -82,21 +82,11 @@ class ShowMyHighlights extends Component<ShowMyHighlightsProps, { showGoToTop: b
       >
         <Filters />
         <Highlights />
-        {this.state.showGoToTop && (
-          <FormattedMessage id='i18n:toolbar:highlights:popup:button:back-to-top'>
-            {(msg: string) => (
-              <Styled.GoToTopWrapper
-                onClick={this.scrollToTop}
-                data-testid='back-to-top-highlights'
-                aria-label={msg}
-              >
-                <Styled.GoToTop>
-                  <Styled.GoToTopIcon />
-                </Styled.GoToTop>
-              </Styled.GoToTopWrapper>
-            )}
-          </FormattedMessage>
-        )}
+        {this.state.showGoToTop && <GoToTopButton
+          i18nAriaLabel='i18n:toolbar:highlights:popup:button:back-to-top'
+          onClick={this.scrollToTop}
+          data-testid='back-to-top-highlights'
+        />}
       </Styled.ShowMyHighlightsBody>
     );
   }
