+++ src/app/content/highlights/components/ShowMyHighlightsStyles.tsx
@@ -1,5 +1,4 @@
 import styled, { css } from 'styled-components/macro';
-import { AngleUp } from 'styled-icons/fa-solid/AngleUp';
 import { labelStyle } from '../../../components/Typography';
 import { h4Style } from '../../../components/Typography/headings';
 import theme from '../../../theme';
@@ -120,36 +119,4 @@ export const HighlightSection = styled.div`
     page-break-after: avoid;
     background: white;
   }
-`;
-
-// tslint:disable-next-line:variable-name
-export const GoToTopWrapper = styled.div`
-  width: 4.8rem;
-  height: 4.8rem;
-  position: absolute;
-  z-index: 2;
-  bottom: 4.8rem;
-  right: 4.8rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-`;
-
-// tslint:disable-next-line:variable-name
-export const GoToTop = styled.div`
-  width: 2.4rem;
-  height: 2.4rem;
-  background: #959595;
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: white;
-`;
-
-// tslint:disable-next-line:variable-name
-export const GoToTopIcon = styled(AngleUp)`
-  width: 1.6rem;
-  height: 1.6rem;
 `;
